requirements.txt
fastapi
uvicorn
tgtg-python
requests
supabase